+++ resources/js/Components/ui/sidebar.tsx
@@ -5,7 +5,7 @@ import { Slot } from "@radix-ui/react-slot"
 import { VariantProps, cva } from "class-variance-authority"
 import { PanelLeft } from "lucide-react"
 
-import { useIsMobile } from "@/hooks/use-mobile"
+import { useIsMobile } from "@/hooks/UseMobile"
 import { cn } from "@/lib/utils"
 import { Button } from "@/Components/ui/button"
 import { Input } from "@/Components/ui/input"
