+++ resources/js/Components/ui/toaster.tsx
@@ -1,4 +1,4 @@
-import { useToast } from "@/hooks/use-toast"
+import { useToast } from "@/hooks/UseToast"
 import {
   Toast,
   ToastClose,
