+++ resources/js/Layouts/AuthenticatedLayout.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from '@/Components/app-sidebar';
+import { AppSidebar } from '@/Components/AppSidebar';
 import {
     Breadcrumb,
     BreadcrumbItem,
