+++ resources/js/Pages/Shipments/Partials/ShipmentNotes.tsx
@@ -12,7 +12,7 @@ import {
     DialogTrigger,
 } from '@/Components/ui/dialog';
 import { Textarea } from '@/Components/ui/textarea';
-import { useToast } from '@/hooks/use-toast';
+import { useToast } from '@/hooks/UseToast';
 import { Note } from '@/types';
 import { Notable } from '@/types/enums';
 import { useForm, usePage } from '@inertiajs/react';
