+++ resources/js/Pages/Shipments/Partials/ShipperDetails.tsx
@@ -2,9 +2,9 @@ import { Badge } from '@/Components/ui/badge';
 import { Button } from '@/Components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/Components/ui/card';
 import { Label } from '@/Components/ui/label';
-import { ResourceSearchSelect } from '@/Components/ui/resource-search-select';
+import { ResourceSearchSelect } from '@/Components/ResourceSearchSelect';
 import { Skeleton } from '@/Components/ui/skeleton';
-import { useToast } from '@/hooks/use-toast';
+import { useToast } from '@/hooks/UseToast';
 import { Shipment } from '@/types';
 import { useForm, usePage } from '@inertiajs/react';
 import { Box, Check, CheckCircle2, Pencil, Truck, X } from 'lucide-react';
